[2014/01/08] modified   SATSepDWTDecoder.h 1. [ADD] Include libclock.h so there is no need to include it before this header.
modified   SepDWTDecoderPool.h
1. [MOD] Cast the type for the template function min().

modified   SepDWTHeader.h
1. [ADD] Define the getters to get the wavelet length, #coef, #coef. within the data domain for each dimension and each level.

+++ WaveletSAT/SepDWTDecoderPool.h
@@ -318,7 +318,9 @@ namespace WaveletSAT
 						iterBins ++ ) 
 					{
 						BT usBin = *iterBins;
-						WT dCumsum = ( !usBin ) ? 0.0 : vFullArray[min(usBin - 1, vFullArray.size()-1)];
+						// MOD-BY-LEETEN 2014/01/08:						WT dCumsum = ( !usBin ) ? 0.0 : vFullArray[min(usBin - 1, vFullArray.size()-1)];
+						WT dCumsum = ( !usBin ) ? 0.0 : vFullArray[min((size_t)usBin - 1, vFullArray.size()-1)];
+						// MOD-BY-LEETEN 2014/01/08-END
 						udValues.push_back(dCumsum);
 					}
 				}
